requirements.txt
openai==0.28
python-dotenv
sentence-transformers
huggingface_hub==0.23.2
scikit-learn